Remove Sender dependance on Policy logic.
The Sender now accepts any value for unscheduledByteLimit from the
Policy::Manager.  Previously, the Sender relied on the Manager to
return an unscheduledByteLimit value less than or equal to the
message's length.

+++ src/Message.h
@@ -18,9 +18,9 @@
 
 #include <Homa/Homa.h>
 
-#include "Protocol.h"
-
 #include <bitset>
+
+#include "Protocol.h"
 
 namespace Homa {
 namespace Core {
@@ -51,7 +51,7 @@ class Message : public virtual Homa::Message {
 
     Driver::Packet* getPacket(uint16_t index) const;
     bool setPacket(uint16_t index, Driver::Packet* packet);
-    uint16_t getNumPackets() const;
+    int getNumPackets() const;
 
     uint32_t rawLength() const;
